first move: re-group aggregates functionalities in core/physical_p/aggregates
