+++ datafusion/core/tests/sql/explain_analyze.rs
@@ -110,7 +110,7 @@ async fn explain_analyze_baseline_metrics() {
         use datafusion::physical_plan::sorts;
 
         plan.as_any().downcast_ref::<sorts::sort::SortExec>().is_some()
-            || plan.as_any().downcast_ref::<physical_plan::hash_aggregate::HashAggregateExec>().is_some()
+            || plan.as_any().downcast_ref::<physical_plan::aggregates::AggregateExec>().is_some()
             // CoalescePartitionsExec doesn't do any work so is not included
             || plan.as_any().downcast_ref::<physical_plan::filter::FilterExec>().is_some()
             || plan.as_any().downcast_ref::<physical_plan::limit::GlobalLimitExec>().is_some()
